# api/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
		"/files/upload": {
			"post": {
				"operationId": "FilesController_upload",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UploadDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": ""
					}
				},
				"tags": [
					"Files"
				]
			}
		},
		"/files/preview/{chatId}/{previewId}": {
			"get": {
				"operationId": "FilesController_downPreview",
				"parameters": [
					{
						"name": "chatId",
						"required": true,
						"in": "path",
						"description": "Chat ID",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "previewId",
						"required": true,
						"in": "path",
						"description": "Preview ID",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Files"
				]
			}
		},
		"/files/{chatId}/{fileId}": {
			"get": {
				"operationId": "FilesController_downFile",
				"parameters": [
					{
						"name": "chatId",
						"required": true,
						"in": "path",
						"description": "Chat ID",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "fileId",
						"required": true,
						"in": "path",
						"description": "File ID",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Returns the file as Buffer",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					}
				},
				"tags": [
					"Files"
				]
			}
		},
		"/files/vpn": {
			"get": {
				"operationId": "FilesController_getVPNProfile",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Files"
				]
			}
		}
	},
	"info": {
		"title": "passim-file-service",
		"description": "",
		"version": "0.0.1",
		"contact": {},
		"license": {
			"name": "UNLICENSED",
			"url": "https://github.com/ramzini22"
		}
	},
	"tags": [],
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "development"
		}
	],
	"components": {
		"schemas": {
			"UploadDto": {
				"type": "object",
				"properties": {
					"chatId": {
						"type": "string"
					}
				},
				"required": [
					"chatId"
				]
			}
		}
	}
}